+++ core/src/main/java/io/questdb/griffin/SqlCompilerImpl.java
@@ -2238,7 +2238,7 @@ public class SqlCompilerImpl implements SqlCompiler, Closeable, SqlParserCallbac
                     if (cursorColumnCount <= writerTimestampIndex) {
                         throw SqlException.$(tableNameExpr.position, "select clause must provide timestamp column");
                     } else {
-                        byte columnType = ColumnType.tagOf(cursorMetadata.getColumnType(writerTimestampIndex));
+                        short columnType = ColumnType.tagOf(cursorMetadata.getColumnType(writerTimestampIndex));
                         if (columnType != ColumnType.TIMESTAMP && columnType != ColumnType.STRING && columnType != ColumnType.NULL) {
                             throw SqlException.$(tableNameExpr.position, "expected timestamp column but type is ").put(ColumnType.nameOf(columnType));
                         }
