+++ core/src/main/java/io/questdb/griffin/engine/functions/rnd/RndStringListFunctionFactory.java
@@ -65,7 +65,7 @@ public class RndStringListFunctionFactory implements FunctionFactory {
         for (int i = 0, n = args.size(); i < n; i++) {
             final Function f = args.getQuick(i);
             if (f.isConstant()) {
-                byte typeTag = ColumnType.tagOf(f.getType());
+                short typeTag = ColumnType.tagOf(f.getType());
                 if (typeTag == ColumnType.STRING || typeTag == ColumnType.NULL) {
                     symbols.add(Chars.toString(f.getStr(null)));
                     continue;
