+++ core/src/test/java/io/questdb/test/cairo/TableWriterTest.java
@@ -4012,7 +4012,7 @@ public class TableWriterTest extends AbstractCairoTest {
             long ts = TimestampFormatUtils.parseTimestamp("2013-03-04T00:00:00.000Z");
 
             Rnd rnd = new Rnd();
-            byte columnTypeTag;
+            short columnTypeTag;
             TableToken tableToken;
             try (TableWriter writer = getWriter(model.getName())) {
